requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
python-docx==0.8.11 